# server/api.yaml
openapi: 3.0.3
info:
  title: Hospital Management System API
  version: 1.0.0
  description: |
    Role-based API for Hospital Management System (HMS).
    Authentication uses JWT Bearer tokens.
servers:
  - url: http://127.0.0.1:5000
    description: Local development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        msg:
          type: string
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        type:
          type: string
          enum: [admin, doctor, patient]
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
...
    Appointment:
      type: object
      properties:
        id:
          type: integer
        doctor_id:
          type: integer
        patient_id:
          type: integer
        doctor_name:
          type: string
          nullable: true
        patient_name:
          type: string
          nullable: true
        start_time:
          type: string
          format: date-time
        end_time:
          type: string
          format: date-time
        status:
          type: string
          enum: [booked, completed, cancelled]
        visit_reason:
          type: string
          nullable: true
        diagnosis:
          type: string
          nullable: true
        prescription:
          type: string
          nullable: true
        tests_requested:
          type: string
          nullable: true
        medications_prescribed:
          type: string
          nullable: true
        follow_up_date:
          type: string
          format: date
          nullable: true
        notes:
          type: string
          nullable: true
paths:
  /api/auth/register:
    post:
      tags: [Auth]
      summary: Register a patient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, name, dob, contact_number]
              properties:
                email:
                  type: string
                password:
                  type: string
                name:
                  type: string
                dob:
                  type: string
                  format: date
                contact_number:
                  type: string
      responses:
        '201':
          description: Patient created
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  refresh_csrf:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
...
  /api/auth/me:
    get:
      tags: [Auth]
      summary: Current user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user
  /api/admin/stats:
    get:
      tags: [Admin]
      summary: Admin dashboard counts
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Counts response
  /api/admin/appointments:
    get:
      tags: [Admin]
      summary: List/search/filter appointments
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: q
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        '200':
          description: Appointments list
          content:
            application/json:
              schema:
                type: object
                properties:
                  appointments:
                    type: array
                    items:
                      $ref: '#/components/schemas/Appointment'
  /api/admin/appointments/{appointment_id}:
    patch:
      tags: [Admin]
      summary: Cancel booked appointment
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: appointment_id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [cancelled]
      responses:
        '200':
          description: Appointment updated
  /api/admin/departments:
    get:
      tags: [Admin]
      summary: List departments
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Department list
    post:
      tags: [Admin]
      summary: Create department
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                description:
                  type: string
      responses:
        '201':
          description: Department created
  /api/admin/doctors:
    get:
      tags: [Admin]
      summary: List doctors
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Doctor list
    post:
      tags: [Admin]
      summary: Create doctor
      security:
        - bearerAuth: []
      responses:
        '201':
          description: Doctor created
  /api/admin/doctors/{doctor_id}:
    get:
      tags: [Admin]
      summary: Get doctor
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: doctor_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Doctor details
    patch:
      tags: [Admin]
      summary: Update doctor
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: doctor_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Doctor updated
    delete:
      tags: [Admin]
      summary: Delete doctor
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: doctor_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Doctor deleted
  /api/admin/patients:
    get:
      tags: [Admin]
      summary: List patients
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Patient list
    post:
      tags: [Admin]
      summary: Create patient
      security:
        - bearerAuth: []
      responses:
        '201':
          description: Patient created
  /api/admin/patients/{patient_id}:
    get:
      tags: [Admin]
      summary: Get patient
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: patient_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Patient details
    patch:
      tags: [Admin]
      summary: Update patient
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: patient_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Patient updated
    delete:
      tags: [Admin]
      summary: Delete patient
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: patient_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Patient deleted
...